Muller: actual initialization of the matrix to zeros
only cont(0,0) was initialized to zero, and since the first line of cont
is not set in the loop below, its elements 1 to 39 were not initialized.

This is likely a neutral operation for most cases since only the first line
of the resulting matrix is affected and only if the initial values given by
the malloc() inside armadillo are not zeroed.

(also move code around so that the track.index(std::string) are performed
only when actually needed)