Moment/SceneApp: fix a subtle bug in setFrameRate()
when the framerate is changed, we *must* recompute the time, not the index
or we could be out of boundaries. For instance, if the new framerate is higher
than the old one and time is close to the end, the computation of index
through time*framerate can be larger than the number of frames.

Another bug that I saw was that when setFrameRate() is called after setPlayerSize(),
the player considers the frameRate to be the old one (default one, i.e. 100)
and therefore scrubbing causes crashes hen going close to the end of the file

A first solution could be to replace all the xxx.setTime(xxx.time(),frameRate)
with xxx.setIndex(xxx.index(),frameRate), but the cleaner solution is to use
setFramerate() and fix the bug there (after all that's what the function is
meant to do)