container: fix compilation issue with g++ 4.8.4 and c++11
compilation error of the type "unsigned int cannot be converted to arma::uword"
when passing lIndex as reference to .min(arma::uword &)

simple fix is to use the return value of .min() instead

note that it was compiling alright with the same compiler, without the
-std=c++11 flag and for all Windows and OSX people, so it's either g++
being overly pedantic or a bug. don't know, don't care.